Build mStat tax table

This function takes a list of taxonomy information and creates a taxonomic table compatible with mStat.

Usage

build_mStat_tax_table(taxlist)

Arguments

taxlist

A list where each element is a named vector representing the taxonomy of a feature.

Value

A matrix representing the taxonomic table.

Examples

# Assume `taxlist` is a list of taxonomy information
# tax_table <- build_mStat_tax_table(taxlist)