Data from the Early Childhood Antibiotics and the Microbiome (ECAM) study

The ecam.obj dataset is derived from a subset of the Early Childhood Antibiotics and the Microbiome (ECAM) study, which tracked the microbiome composition and development of 43 infants in the United States from birth to 2 years of age, identifying associations between the microbiome and antibiotic exposure, delivery mode, and diet.

Usage

ecam.obj

Format

An object of class list of length 3.

Source

Data source: https://github.com/FrederickHuangLin/ANCOM/tree/master/data

Details

ecam.obj is a MicrobiomeStat Data Object.

References

Bokulich, Nicholas A., et al. "Antibiotics, birth mode, and diet shape microbiome maturation during early life." Science translational medicine 8.343 (2016): 343ra82-343ra82.