Generate Beta Ordination Plots

This function generates beta ordination plots using the specified distance measure. The plots can be stratified by a given variable and the color and shape aesthetics can be mapped to groups and time variables, respectively. The function also supports saving the plots as PDF files.

Usage

generate_beta_ordination_long(
  data.obj = NULL,
  dist.obj = NULL,
  pc.obj = NULL,
  subject.var,
  time.var,
  t0.level,
  ts.levels,
  group.var = NULL,
  strata.var = NULL,
  dist.name = c("BC", "Jaccard"),
  base.size = 16,
  theme.choice = "prism",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)

Arguments

data.obj

a list with the data to be used for the plots. If NULL, the distance object will be used instead.

dist.obj

a list with distance matrices. If NULL, it will be calculated from the data object.

pc.obj

a list with principal coordinates. If NULL, it will be calculated from the distance object.

subject.var

a character string specifying the subject variable.

time.var

a character string specifying the time variable.

t0.level

a character string specifying the baseline time level.

ts.levels

a vector of character strings specifying the time series levels.

group.var

a character string specifying the grouping variable. If NULL, only color will be used in the plots.

strata.var

a character string specifying the stratification variable. If NULL, no stratification will be done.

dist.name

a character vector specifying the distance measures to use. Defaults to c('BC', 'Jaccard').

base.size

a numeric value specifying the base size for the plot text.

theme.choice

a character string specifying the theme to use. Defaults to "prism".

custom.theme

a ggplot2 theme object to use instead of the default theme.

palette

a character vector specifying the colors to use for the plots. If NULL, a default palette will be used.

pdf

a logical value indicating whether to save the plots as PDF files. Defaults to TRUE.

file.ann

a character string specifying an annotation to add to the file names of the saved plots.

pdf.wid

a numeric value specifying the width of the saved PDF files.

pdf.hei

a numeric value specifying the height of the saved PDF files.

...

further arguments to be passed to the underlying functions.

Value

A list of ggplot2 objects representing the beta ordination plots.

See also

Author

Caffery Yang cafferychen7850@gmail.com

Examples

if (FALSE) {
data(subset_T2D.obj)
generate_beta_ordination_long(
  data.obj = subset_T2D.obj,
  dist.obj = NULL,
  pc.obj = NULL,
  subject.var = "subject_id",
  time.var = "visit_number",
  t0.level = sort(unique(subset_T2D.obj$meta.dat$visit_number))[1],
  ts.levels = sort(unique(subset_T2D.obj$meta.dat$visit_number))[2:6],
  group.var = "subject_race",
  strata.var = "subject_gender",
  dist.name = c("BC"),
  base.size = 16,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5
)
}