Generate Beta Diversity Ordination Plot

This function generates ordination plots (Principal Component Analysis) based on beta-diversity distances. It also allows for stratification and grouping of samples, and calculation of distances at a specific time point.

Usage

generate_beta_ordination_single(
  data.obj = NULL,
  dist.obj = NULL,
  pc.obj = NULL,
  subject.var,
  time.var = NULL,
  t.level = NULL,
  group.var = NULL,
  strata.var = NULL,
  dist.name = c("BC", "Jaccard"),
  base.size = 16,
  theme.choice = "prism",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)

Arguments

data.obj

A data object. Default is NULL.

dist.obj

A distance object. Default is NULL.

pc.obj

A principal component object. Default is NULL.

subject.var

String. Variable to be used as subject.

time.var

String. Variable to be used for time. Default is NULL.

t.level

Time level to filter the data by. Default is NULL.

group.var

String. Variable to be used for grouping. Default is NULL.

strata.var

String. Variable to be used for stratification. Default is NULL.

dist.name

Character vector. Name of distance(s) to be used. Default is c('BC', 'Jaccard').

base.size

Numeric. Base size for plot elements. Default is 16.

theme.choice

String. Theme choice for the plot. Default is 'prism'.

custom.theme

A custom theme to be used. Default is NULL.

palette

A custom color palette to be used. Default is NULL.

pdf

Logical. If TRUE, the plots are saved as PDF files. Default is TRUE.

file.ann

File annotation. Default is NULL.

pdf.wid

Width of the PDF. Default is 11.

pdf.hei

Height of the PDF. Default is 8.5.

...

Additional arguments to be passed.

Value

A PCoA plot displaying the beta diversity ordination, stratified by the specified grouping and/or strata variables (if provided). The plot will be saved as a PDF if `pdf` is set to `TRUE`.

Details

The function is flexible and allows for various modifications, including the choice of distance measure and stratification factor, providing a comprehensive tool for microbiome beta diversity exploration. It integrates well with other MicrobiomeStat functions and takes their output as input.

See also

Author

Caffery Yang cafferychen7850@gmail.com

Examples

if (FALSE) {
# Load required libraries and example data
library(vegan)
data(peerj32.obj)
dist.obj <- mStat_calculate_beta_diversity(peerj32.obj, dist.name = c('BC', 'Jaccard'))
pc.obj <- mStat_calculate_PC(dist.obj, method = c('mds'), k = 2, dist.name = c('BC','Jaccard'))
# Generate the boxplot pair
generate_beta_ordination_single(
  data.obj = peerj32.obj,
  dist.obj = dist.obj,
  pc.obj = pc.obj,
  subject.var = "subject",
  time.var = "time",
  t.level = "2",
  group.var = "group",
  strata.var = "sex",
  dist.name = c("BC"),
  base.size = 20,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = "test",
  pdf.wid = 11,
  pdf.hei = 8.5
)
}