Beta Diversity Change Boxplot Pairs

This function generates boxplots to visualize the changes in beta diversity principal components (PCs) over time. It allows the use of Principal Coordinates Analysis (PCoA), non-metric multidimensional scaling (NMDS), t-SNE, or UMAP for dimension reduction.

Usage

generate_beta_pc_change_boxplot_pair(
  data.obj = NULL,
  dist.obj = NULL,
  pc.obj = NULL,
  pc.ind = c(1, 2),
  subject.var,
  time.var,
  group.var = NULL,
  strata.var = NULL,
  change.base = NULL,
  change.func = "difference",
  dist.name = c("BC", "Jaccard"),
  base.size = 16,
  theme.choice = "prism",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)

Arguments

data.obj

A list containing the input data. Default is NULL.

dist.obj

A list containing the distance object. Default is NULL.

pc.obj

A list containing the Principal Component Analysis object. Default is NULL.

pc.ind

A numeric vector indicating the PC indexes used. Default is c(1, 2).

subject.var

A string specifying the variable for subjects.

time.var

A string specifying the variable for time.

group.var

A string specifying the variable for groups. Default is NULL.

strata.var

A string specifying the variable for strata. Default is NULL.

change.base

The baseline for calculating changes in beta diversity. Default is NULL.

change.func

A function or string specifying how to calculate changes. Default is "difference".

dist.name

A character vector indicating the distance metrics used. Default is c("BC", "Jaccard").

base.size

A numeric value for the base size of the plot. Default is 16.

theme.choice

A string specifying the theme of the plot. Default is "prism".

custom.theme

A ggplot2 theme object for user-defined theme. Default is NULL.

palette

A character vector specifying the color palette. Default is NULL.

pdf

A logical value indicating whether to save the plot as a PDF. Default is TRUE.

file.ann

A string for additional annotation to the file name. Default is NULL.

pdf.wid

A numeric value specifying the width of the PDF. Default is 11.

pdf.hei

A numeric value specifying the height of the PDF. Default is 8.5.

...

Additional arguments to be passed to the function.

Value

A list of ggplot objects for each PC index and distance metric.

Details

This function generates a boxplot of changes in beta diversity based on PCoA coordinates for longitudinal data. The boxplot can be stratified by a group variable and/or other variables. It also allows for different distance metrics and principal component indexes to be used. The function can handle a large number of time points or subjects by averaging the data and adding jitter to the plot. The function also has options to customize the size, theme, and color palette of the plot, and to save the plot as a PDF.

Examples

if (FALSE) {
# Load required libraries and example data
library(vegan)
library(ggh4x)
data(peerj32.obj)

# Generate the boxplot pair
generate_beta_pc_change_boxplot_pair(
  data.obj = peerj32.obj,
  dist.obj = NULL,
  pc.obj = NULL,
  pc.ind = c(1, 2),
  subject.var = "subject",
  time.var = "time",
  group.var = "group",
  strata.var = NULL,
  change.base = "1",
  change.func = "difference",
  dist.name = c('BC'),
  base.size = 20,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5
)
}