Generate Individual Taxa Change Boxplot Pair

This function generates boxplot pairs visualizing the changes in abundance of individual taxa over time. The boxplots show the change in abundance for each taxon for different groups and strata. It also allows for a prevalence and abundance filter to be applied to the data, and can optionally save the plots as a PDF.

Usage

generate_taxa_change_boxplot_pair(
  data.obj,
  subject.var,
  time.var,
  group.var = NULL,
  strata.var = NULL,
  change.base = NULL,
  change.func = "relative difference",
  feature.level = NULL,
  feature.dat.type = c("count", "proportion", "other"),
  features.plot = NULL,
  top.k.plot = NULL,
  top.k.func = NULL,
  prev.filter,
  abund.filter,
  base.size = 16,
  theme.choice = "prism",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)

Arguments

data.obj

A data object containing the data to be plotted.

subject.var

A character string specifying the subject variable.

time.var

A character string specifying the time variable.

group.var

A character string specifying the group variable. Default is NULL.

strata.var

A character string specifying the strata variable. Default is NULL.

change.base

The time point to be used as the base for calculating change in abundance.

change.func

The function to be used for calculating change in abundance. Default is "relative difference".

feature.level

The feature level at which to plot the data.

feature.dat.type

The type of the feature data. Can be "count", "proportion", or "other". Default is "count".

features.plot

A character vector specifying the features to be plotted. Default is NULL.

top.k.plot

The top k features to be plotted. Default is NULL.

top.k.func

The function to be used for selecting the top k features. Default is NULL.

prev.filter

The prevalence filter to be applied to the data.

abund.filter

The abundance filter to be applied to the data.

base.size

The base size for the plot. Default is 16.

theme.choice

The theme choice for the plot. Default is "prism".

custom.theme

An optional custom theme for the plot. Default is NULL.

palette

A character vector specifying the color palette for the plot. Default is NULL.

pdf

A logical value indicating whether to save the plot as a PDF. Default is TRUE.

file.ann

An optional character string to be appended to the file name of the PDF. Default is NULL.

pdf.wid

The width of the PDF. Default is 11.

pdf.hei

The height of the PDF. Default is 8.5.

...

Additional arguments passed to the underlying functions.

Value

A list of ggplot objects, each of which is a boxplot visualizing the changes in abundance of individual taxa over time.

Examples

if (FALSE) {
# Load required libraries and data
library(vegan)
library(ggh4x)
data(peerj32.obj)

# Generate the boxplot pair
generate_taxa_change_boxplot_pair(
  data.obj = peerj32.obj,
  subject.var = "subject",
  time.var = "time",
  group.var = "group",
  strata.var = NULL,
  change.base = "1",
  change.func = "lfc",
  feature.level = c("Family"),
  feature.dat.type = "count",
  features.plot = NULL,
  top.k.plot = 6,
  top.k.func = "sd",
  prev.filter = 0.01,
  abund.filter = 0.01,
  base.size = 20,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = "test",
  pdf.wid = 11,
  pdf.hei = 8.5
)
}