Generate Taxonomic Dotplot Single

This function generates a stacked dotplot of specified taxa level with paired samples. The data used in this visualization will be first filtered based on prevalence and abundance thresholds. The plot can either be displayed interactively or saved as a PDF file.

Usage

generate_taxa_dotplot_single(
  data.obj,
  subject.var,
  time.var = NULL,
  t.level = NULL,
  group.var = NULL,
  strata.var = NULL,
  feature.level = NULL,
  feature.dat.type = c("count", "proportion", "other"),
  features.plot = NULL,
  top.k.plot = NULL,
  top.k.func = NULL,
  prev.filter = 0.01,
  abund.filter = 0.01,
  base.size = 16,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)

Arguments

data.obj

A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list).

subject.var

A character string defining subject variable in meta_tab

time.var

A character string defining time variable in meta_tab

t.level

The base level for time points in longitudinal data.

group.var

A character string defining group variable in meta_tab used for sorting and facetting

strata.var

(Optional) A character string defining strata variable in meta_tab used for sorting and facetting

feature.level

A character string defining the taxonomic level to analyze ('Phylum', 'Family' or 'Genus')

feature.dat.type

A character string specifying the type of data in feature.dat. Options are "count", "proportion", or "other".

features.plot

A character vector specifying the taxa to be plotted. If NULL (default), the top k taxa by mean abundance will be plotted.

top.k.plot

A numeric value specifying the number of top taxa to be plotted if features.plot is NULL. If NULL (default), all taxa will be plotted.

top.k.func

A function to compute the top k taxa if features.plot is NULL. If NULL (default), the mean function will be used.

prev.filter

A numeric value defining the prevalence threshold to filter taxa, between 0 and 1.

abund.filter

A numeric value defining the abundance threshold to filter taxa.

base.size

Base font size for the generated plots.

theme.choice

Plot theme choice (default: "bw").

custom.theme

Custom ggplot2 theme (optional).

palette

Color palette used for the plots.

pdf

If TRUE, save the plot as a PDF file (default: TRUE)

file.ann

The file name annotation (default: NULL)

pdf.wid

Width of the PDF plots.

pdf.hei

Height of the PDF plots.

...

Additional parameters to be passed

Value

If the `pdf` parameter is set to TRUE, the function will save a PDF file and return the final ggplot object. If `pdf` is set to FALSE, the function will return the final ggplot object without creating a PDF file.

Examples

if (FALSE) {
# Load required libraries
library(vegan)
library(ggh4x)
data(peerj32.obj)

# Call the function
generate_taxa_dotplot_single(
  data.obj = peerj32.obj,
  subject.var = "subject",
  time.var = NULL,
  t.level = NULL,
  group.var = "group",
  strata.var = NULL,
  feature.level = c("Family"),
  feature.dat.type = "count",
  features.plot = NULL,
  top.k.plot = NULL,
  top.k.func = NULL,
  prev.filter = 0.01,
  abund.filter = 0.01,
  base.size = 16,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 15,
  pdf.hei = 8.5
)
}