Conduct Differential Abundance Testing Using ZicoSeq Method in MicrobiomeStat Package

This function applies a differential abundance analysis using ZicoSeq on a data set. The function filters taxa based on prevalence and abundance, then it aggregates and applies the ZicoSeq method. Finally, it creates a report of significant taxa with relevant statistics.

Usage

generate_taxa_test_single(
  data.obj,
  time.var = NULL,
  t.level = NULL,
  group.var,
  adj.vars,
  prev.filter = 0,
  abund.filter = 0,
  feature.level,
  feature.dat.type = c("count", "proportion", "other"),
  ...
)

Arguments

data.obj

A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.

time.var

Character string specifying the column name in metadata containing time variable. Used to subset data to a single timepoint if provided. Default NULL does not subset.

t.level

Character string specifying the time level/value to subset data to, if a time variable is provided. Default NULL does not subset data.

group.var

Character string specifying the column name in metadata containing grouping categories. This will be used as the predictor in differential abundance testing.

adj.vars

Character vector specifying column names in metadata containing covariates. These will be used for adjustment in differential abundance testing.

prev.filter

Numeric value specifying the minimum prevalence threshold for filtering taxa before analysis. Taxa with prevalence below this value will be removed. Prevalence is calculated as the proportion of samples where the taxon is present. Default 0 removes no taxa by prevalence filtering.

abund.filter

Numeric value specifying the minimum abundance threshold for filtering taxa before analysis. Taxa with mean abundance below this value will be removed. Abundance refers to counts or proportions depending on feature.dat.type. Default 0 removes no taxa by abundance filtering.

feature.level

The column name in the feature annotation matrix (feature.ann) of data.obj to use for summarization and plotting. This can be the taxonomic level like "Phylum", or any other annotation columns like "Genus" or "OTU_ID". Should be a character vector specifying one or more column names in feature.ann. Multiple columns can be provided, and data will be plotted separately for each column. Default is NULL, which defaults to all columns in feature.ann if `features.plot` is also NULL.

feature.dat.type

The type of the feature data, which determines how the data is handled in downstream analyses. Should be one of: - "count": Raw count data, will be normalized by the function. - "proportion": Data that has already been normalized to proportions/percentages. - "other": Custom abundance data that has unknown scaling. No normalization applied. The choice affects preprocessing steps as well as plot axis labels. Default is "count", which assumes raw OTU table input.

...

Additional arguments to be passed to the ZicoSeq function.

Value

A list of tibble(s) containing information about significant taxa, including R.Squared, F.Statistic, Estimate, P.Value, Adjusted.P.Value, Mean.Proportion, Mean.Prevalence, SD.Abundance and SD.Prevalence.

Details

This function performs differential abundance analysis using ZicoSeq method:

1. Subset data to specific time point if time variable and level provided.

2. Extract OTU table, taxonomy table, and sample metadata.

3. Filter OTUs by prevalence and abundance thresholds if counting data.

4. Aggregate OTU table to taxonomic levels specified by feature.level.

5. Run ZicoSeq on aggregated table, with grouping and adjustment variables.

6. Extract and compile statistics for significant taxa into results tables, including R-squared, F statistic, coefficients, p-values, mean proportions, etc.

7. Return list of tables, with each element corresponding to one taxonomic level.

In summary, it applies preprocessing steps, runs ZicoSeq differential abundance testing, and compiles informative results tables for significant taxa. Adjustment for covariates is supported. Customizable taxonomic aggregation allows testing at different levels.

Examples

if (FALSE) {
data(peerj32.obj)
generate_taxa_test_single(
    data.obj = peerj32.obj,
    time.var = "time",
    t.level = "1",
    group.var = "group",
    adj.vars = "sex",
    feature.dat.type = "count",
    feature.level = "Genus",
    prev.filter = 0,
    abund.filter = 0,
    is.winsor = TRUE,
    outlier.pct = 0.001,
    winsor.end = 'top',
    is.post.sample = TRUE,
    post.sample.no = 25,
    list(function (x) x^0.5, function (x) x^0.25),
    stats.combine.func = max,
    perm.no = 99,
    strata = NULL,
    ref.pct = 0.5,
    stage.no = 6,
    excl.pct = 0.2,
    is.fwer = TRUE,
    verbose = TRUE
)
}