Generate volcano plots for taxa differential test for a single time point

Generate volcano plots for taxa differential test for a single time point

Usage

generate_taxa_volcano_single(
  data.obj,
  group.var = NULL,
  test.list,
  feature.sig.level = 0.1,
  feature.mt.method = "fdr"
)

Arguments

data.obj

A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.

group.var

The grouping variable tested, found in metadata

test.list

The list of test results returned by generate_taxa_trend_test_long

feature.sig.level

The significance level cutoff for highlighting taxa

feature.mt.method

Multiple testing correction method, "fdr" or "none"

Value

A list of ggplot objects of volcano plots for each taxonomic level

Examples

data(peerj32.obj)
test.list <- generate_taxa_test_single(
    data.obj = peerj32.obj,
    time.var = "time",
    t.level = "2",
    group.var = "group",
    adj.vars = "sex",
    feature.dat.type = "count",
    feature.level = c("Phylum","Genus","Family"),
    prev.filter = 0.1,
    abund.filter = 0.0001,
    feature.sig.level = 0.1,
    feature.mt.method = "none"
)
#> Rule 1 passed: data.obj is a list.
#> Rule 2 passed: meta.dat is a data.frame.
#> Rule 3 passed: The row names of feature.tab match the row names of feature.ann.
#> Rule 4 passed: The column names of feature.tab match the row names of meta.dat.
#> Rule 5 passed: feature.tab is a matrix.
#> Rule 6 passed: feature.ann is a matrix.
#> Please note: The data components should follow base R data.frame and matrix structures, not phyloseq's formal class.
#> Validation passed.
#> Note: Passing validation does not guarantee the absence of all data issues. Further data exploration may be needed.
#> Data has been subsetted based on the provided condition.
#> Updated metadata to match the subsetted data.
#> The following samples were excluded: sample-1, sample-3, sample-5, sample-7, sample-9, sample-11, sample-13, sample-15, sample-17, sample-19, sample-21, sample-23, sample-25, sample-27, sample-29, sample-31, sample-33, sample-35, sample-37, sample-39, sample-41, sample-43
#> Updated feature table to match the subsetted data.
#> Updated feature annotation to match the subsetted data.
#> Data subsetting complete. Returning updated data object.
#> Your data is in raw format ('Raw'). Normalization is crucial for further analyses. Now, 'mStat_normalize_data' function is automatically applying 'TSS' transformation.
#> Data has been successfully normalized using TSS method.
#> 0 features are filtered!
#> The filtered data has 22 samples and 6 features that will be tested!
#> Fit linear models ...
#> Completed.
#> 0 features are filtered!
#> The filtered data has 22 samples and 115 features that will be tested!
#> Fit linear models ...
#> Completed.
#> 0 features are filtered!
#> The filtered data has 22 samples and 18 features that will be tested!
#> Fit linear models ...
#> Completed.
   volcano_plots <- generate_taxa_volcano_single(data.obj = peerj32.obj,
                                                 group.var = "group",
                                                 test.list = test.list,
                                                 feature.sig.level = 0.1,
                                                 feature.mt.method = "none")