Generate a Comprehensive Report for Microbial Ecology Analysis of Longitudinal Data

This function generates a comprehensive report for microbial ecology analysis, including changes in alpha diversity, beta diversity, and taxonomic features in longitudinal data.

Usage

mStat_generate_report_long(
  data.obj,
  dist.obj = NULL,
  alpha.obj = NULL,
  depth = NULL,
  group.var,
  adj.vars,
  subject.var,
  time.var,
  alpha.name = c("shannon", "simpson", "observed_species", "chao1", "ace", "pielou"),
  dist.name = c("BC", "Jaccard", "UniFrac", "GUniFrac", "WUniFrac", "JS"),
  t0.level,
  ts.levels,
  strata.var = NULL,
  base.size = 16,
  theme.choice = "prism",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  prev.filter = 0,
  abund.filter = 0,
  feature.level = NULL,
  feature.dat.type = c("count", "proportion", "other"),
  Transform = c("identity", "sqrt", "log"),
  output.file,
  ...
)

Arguments

data.obj

A data object created by mStat_convert_phyloseq_to_data_obj.

dist.obj

A distance object created by mStat_calculate_beta_diversity.

alpha.obj

An alpha diversity object (optional).

depth

Sampling depth (optional).

group.var

Variable name used for grouping samples.

adj.vars

Variables to adjust for in the analysis.

subject.var

Variable name used for subject identification.

time.var

Variable name used for time points in longitudinal data.

alpha.name

Names of alpha diversity indices to include in the analysis.

dist.name

Names of beta diversity distance metrics to include in the analysis.

t0.level

The base level for time points in longitudinal data.

ts.levels

The levels for time points in longitudinal data.

strata.var

Variable to stratify the analysis by (optional).

base.size

Base font size for the generated plots.

theme.choice

Plot theme choice (default: "prism").

custom.theme

Custom ggplot2 theme (optional).

palette

Color palette used for the plots.

pdf

Logical indicating whether to save plots as PDF files (default: TRUE).

file.ann

Annotation text for the PDF file names.

pdf.wid

Width of the PDF plots.

pdf.hei

Height of the PDF plots.

prev.filter

Prevalence filter for feature analysis.

abund.filter

Abundance filter for feature analysis.

feature.level

Taxonomic level for feature analysis.

feature.dat.type

Data type for feature analysis (count, proportion, or other).

Transform

The transformation function applied to the data (default: "log").

output.file

Output file name for the report.

...

Additional arguments passed to internal functions.

Value

A PDF report file containing the microbial ecology analysis results for longitudinal data.

Examples

if (FALSE) {
# Assuming peerj32.obj, dist.obj, alpha.obj are pre-defined objects
data(peerj32.obj)
mStat_generate_report_long(
  data.obj = peerj32.obj,
  dist.obj = NULL,
  alpha.obj = NULL,
  group.var = "group",
  adj.vars = c("sex"),
  subject.var = "subject",
  time.var = "time",
  alpha.name = c("shannon","simpson"),
  dist.name = c("BC",'Jaccard'),
  t0.level = "1",
  ts.levels = "2",
  strata.var = "sex",
  feature.level = c("Phylum"),
  feature.dat.type = "count",
  Transform = "log",
  theme.choice = "bw",
  base.size = 12,
  output.file = "report.pdf"
)
}