Convert Mothur Data into MicrobiomeStat's Data Object

This function serves as a bridge between Mothur and MicrobiomeStat, allowing seamless integration of Mothur data with the powerful MicrobiomeStat analysis toolkit. It accepts various Mothur microbiome study-related files and converts them into a MicrobiomeStat data object.

Usage

mStat_import_mothur_as_data_obj(
  mothur_list_file = NULL,
  mothur_group_file = NULL,
  mothur_tree_file = NULL,
  cutoff = NULL,
  mothur_shared_file = NULL,
  mothur_constaxonomy_file = NULL,
  parseFunction = parse_taxonomy_default
)

Arguments

mothur_list_file

(Optional) A list file from Mothur.

mothur_group_file

(Optional) A group file from Mothur.

mothur_tree_file

(Optional) A tree file from Mothur.

cutoff

(Optional) A cutoff value for OTU.

mothur_shared_file

(Optional) A shared file from Mothur.

mothur_constaxonomy_file

(Optional) A constaxonomy file from Mothur.

parseFunction

(Optional) A function for parsing taxonomy. Defaults to parse_taxonomy_default.

Value

A MicrobiomeStat data object that contains the feature table, metadata, feature annotations, phylogenetic tree, and normalization status. This object is ready to be used in MicrobiomeStat's downstream analysis functions.

Details

MicrobiomeStat is a comprehensive and user-friendly tool designed for robust microbiome data analysis. The resulting data object from this function can be directly fed into a series of MicrobiomeStat's statistical tools for further analysis.

See also

Author

Caffery Yang

Examples

if (FALSE) { # \dontrun{
# Make sure to load the plyr package
# library(plyr)
# path_to_list_file <- "/Users/apple/Downloads/esophagus/Esophagus/esophagus.fn.list"
# path_to_group_file <- "/Users/apple/Downloads/esophagus/Esophagus/esophagus.good.groups"
# path_to_tree_file <- "/Users/apple/Downloads/esophagus/Esophagus/esophagus.tree"
# path_to_shared_file <- "/Users/apple/Downloads/esophagus/Esophagus/esophagus.fn.shared"

# data_obj <- mStat_import_mothur_as_data_obj(mothur_list_file = path_to_list_file,
                                           # mothur_group_file = path_to_group_file,
                                           # mothur_tree_file = path_to_tree_file,
                                           # mothur_shared_file = path_to_shared_file
                                           # )
} # }