Data from the Early Childhood Antibiotics and the Microbiome (ECAM) study

The ecam.obj dataset is derived from a subset of the Early Childhood Antibiotics and the Microbiome (ECAM) study, which tracked the microbiome composition and development of 43 infants in the United States from birth to 2 years of age, identifying associations between the microbiome and antibiotic exposure, delivery mode, and diet.

Usage

ecam.obj

Format

An object of class list of length 3.

Source

Data source: https://github.com/FrederickHuangLin/ANCOM/tree/master/data

Details

ecam.obj is a MicrobiomeStat Data Object.

The meta.dat component of ecam.obj is a data frame with 875 observations and 14 variables:

X.SampleID: Character. Unique identifier for each sample.

antiexposedall: Character. Indicates antibiotic exposure ('y' or 'n').

day_of_life: Character. Day of life when the sample was collected.

delivery: Character. Mode of delivery (e.g., 'Vaginal').

diet: Character. Diet type (e.g., 'bd' for breast-fed).

diet_3: Character. Categorized diet type (e.g., 'eb' for exclusively breast-fed).

mom_child: Character. Indicates whether the sample is from mother or child ('C' for child).

month: Character. Month when the sample was collected.

month_of_life: Character. Precise month of life when the sample was collected.

sample_summary: Character. Summary of sample characteristics.

sex: Character. Sex of the infant.

studyid: Character. Study identifier.

subject.id: Character. Unique identifier for each subject.

month_num: Numeric. Month number.

References

Bokulich, Nicholas A., et al. "Antibiotics, birth mode, and diet shape microbiome maturation during early life." Science translational medicine 8.343 (2016): 343ra82-343ra82.