Generate beta diversity change spaghetti plot for longitudinal data

This function generates a spaghetti plot visualizing the change in beta diversity over time for different groups and strata in longitudinal data. Optionally, a user can specify whether to save the plot as a PDF file. Change is defined as distance between samples from two time points.

Usage

generate_beta_change_spaghettiplot_long(
  data.obj,
  dist.obj = NULL,
  subject.var,
  time.var,
  t0.level = NULL,
  ts.levels = NULL,
  group.var = NULL,
  strata.var = NULL,
  adj.vars = NULL,
  dist.name = c("BC"),
  base.size = 16,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)

Arguments

data.obj

A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.

dist.obj

Distance matrix between samples, usually calculated using mStat_calculate_beta_diversity function. If NULL, beta diversity will be automatically computed from data.obj using mStat_calculate_beta_diversity.

subject.var

The variable in the metadata table that represents the subject.

time.var

The variable in the metadata table that represents the time.

t0.level

Character or numeric, baseline time point for longitudinal analysis, e.g. "week_0" or 0. Required.

ts.levels

Character vector, names of follow-up time points, e.g. c("week_4", "week_8"). Required.

group.var

(Optional) The variable in the metadata table that represents the grouping factor.

strata.var

(Optional) The variable in the metadata table that represents the stratification factor.

adj.vars

A character vector containing the names of the columns in data.obj$meta.dat to include as covariates in the PERMANOVA analysis. If no covariates are needed, use NULL (default).

dist.name

A character vector specifying which beta diversity indices to calculate. Supported indices are "BC" (Bray-Curtis), "Jaccard", "UniFrac" (unweighted UniFrac), "GUniFrac" (generalized UniFrac), "WUniFrac" (weighted UniFrac), and "JS" (Jensen-Shannon divergence). If a name is provided but the corresponding object does not exist within dist.obj, it will be computed internally. If the specific index is not supported, an error message will be returned.

base.size

(Optional) Base font size for the plot (default is 16).

theme.choice

Plot theme choice. Can be one of: - "prism": ggprism::theme_prism() - "classic": theme_classic() - "gray": theme_gray() - "bw": theme_bw() Default is "bw".

custom.theme

A custom ggplot theme provided as a ggplot2 theme object. This allows users to override the default theme and provide their own theme for plotting. To use a custom theme, first create a theme object with ggplot2::theme(), then pass it to this argument. For example:

```r my_theme <- ggplot2::theme( axis.title = ggplot2::element_text(size=16, color="red"), legend.position = "none" ) ```

Then pass `my_theme` to `custom.theme`. Default is NULL, which will use the default theme based on `theme.choice`.

palette

(Optional) A palette function or character vector with the colors for the plot.

pdf

(Optional) A boolean indicating whether to save the output as a PDF file (default is TRUE).

file.ann

(Optional) A string for annotating the output file name.

pdf.wid

(Optional) The width of the PDF file if `pdf` is set to `TRUE` (default is 11).

pdf.hei

(Optional) The height of the PDF file if `pdf` is set to `TRUE` (default is 8.5).

...

(Optional) Additional arguments to pass to the plotting function.

Value

A spaghetti plot displaying the beta diversity change over time, stratified by the specified grouping and/or strata variables (if provided). The plot will be saved as a PDF if `pdf` is set to `TRUE`.

Details

This function calculates beta diversity distances between all pairs of samples using the specified distance metric (dist_name). The change in beta diversity for each subject is then defined as the distance between the baseline sample (t0_level) and the sample at each follow-up timepoint (ts_levels). These changes are visualized using spaghetti lines connecting the distances over time.

Examples

if (FALSE) {
data(ecam.obj)
generate_beta_change_spaghettiplot_long(
  data.obj = ecam.obj,
  dist.obj = NULL,
  subject.var = "studyid",
  time.var = "month",
  t0.level = NULL,
  ts.levels = NULL,
  group.var = "diet",
  strata.var = "sex",
  adj.vars = NULL,
  dist.name = c("BC"),
  base.size = 20,
  theme.choice = "bw",
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5
)
}