Generate Beta Ordination Plots

This function generates beta ordination plots using the specified distance measure. The plots can be stratified by a given variable and the color and shape aesthetics can be mapped to groups and time variables, respectively. The function also supports saving the plots as PDF files.

Usage

generate_beta_ordination_long(
  data.obj = NULL,
  dist.obj = NULL,
  pc.obj = NULL,
  subject.var,
  time.var,
  t0.level = NULL,
  ts.levels = NULL,
  group.var = NULL,
  strata.var = NULL,
  adj.vars = NULL,
  dist.name = c("BC", "Jaccard"),
  base.size = 16,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)

Arguments

data.obj

A list object in a format specific to MicrobiomeStat, which can include components such as feature.tab (matrix), feature.ann (matrix), meta.dat (data.frame), tree, and feature.agg.list (list). The data.obj can be converted from other formats using several functions from the MicrobiomeStat package, including: 'mStat_convert_DGEList_to_data_obj', 'mStat_convert_DESeqDataSet_to_data_obj', 'mStat_convert_phyloseq_to_data_obj', 'mStat_convert_SummarizedExperiment_to_data_obj', 'mStat_import_qiime2_as_data_obj', 'mStat_import_mothur_as_data_obj', 'mStat_import_dada2_as_data_obj', and 'mStat_import_biom_as_data_obj'. Alternatively, users can construct their own data.obj. Note that not all components of data.obj may be required for all functions in the MicrobiomeStat package.

dist.obj

Distance matrix between samples, usually calculated using mStat_calculate_beta_diversity function. If NULL, beta diversity will be automatically computed from data.obj using mStat_calculate_beta_diversity.

pc.obj

A list containing the results of dimension reduction/Principal Component Analysis. This should be the output from functions like mStat_calculate_PC, containing the PC coordinates and other metadata. If NULL (default), dimension reduction will be automatically performed using metric multidimensional scaling (MDS) via mStat_calculate_PC. The pc.obj list structure should contain:

$points:A matrix with samples as rows and PCs as columns containing the coordinates.

$eig:Eigenvalues for each PC dimension.

$vectors:Loadings vectors for features onto each PC.

Other metadata like $method, $dist.name, etc.

See mStat_calculate_PC function for details on output format.

subject.var

Character string specifying the column name in metadata containing the subject IDs. This should uniquely identify each subject in the study. Required for connecting samples from the same subject.

time.var

Character string specifying the column name in metadata containing the time variable. This should contain time points for each sample. Required for ordering and connecting samples for the same subject over time.

t0.level

Character or numeric, baseline time point for longitudinal analysis, e.g. "week_0" or 0. Required.

ts.levels

Character vector, names of follow-up time points, e.g. c("week_4", "week_8"). Required.

group.var

Character string specifying the column name in metadata containing the grouping variable. This will be mapped to color aesthetic in the plot. Optional, can be NULL.

strata.var

Character string specifying the column name in metadata containing the stratification variable. This will be used for nested faceting in the plots. Optional, can be NULL.

adj.vars

Character vector specifying column names in metadata to include as covariates in multivariate adjustment of the distance matrix before ordination. Can be empty or NULL if no adjustment needed.

dist.name

A character vector specifying which beta diversity indices to calculate. Supported indices are "BC" (Bray-Curtis), "Jaccard", "UniFrac" (unweighted UniFrac), "GUniFrac" (generalized UniFrac), "WUniFrac" (weighted UniFrac), and "JS" (Jensen-Shannon divergence). If a name is provided but the corresponding object does not exist within dist.obj, it will be computed internally. If the specific index is not supported, an error message will be returned.

base.size

a numeric value specifying the base size for the plot text.

theme.choice

Plot theme choice. Can be one of: - "prism": ggprism::theme_prism() - "classic": theme_classic() - "gray": theme_gray() - "bw": theme_bw() Default is "bw".

custom.theme

A custom ggplot theme provided as a ggplot2 theme object. This allows users to override the default theme and provide their own theme for plotting. To use a custom theme, first create a theme object with ggplot2::theme(), then pass it to this argument. For example:

```r my_theme <- ggplot2::theme( axis.title = ggplot2::element_text(size=16, color="red"), legend.position = "none" ) ```

Then pass `my_theme` to `custom.theme`. Default is NULL, which will use the default theme based on `theme.choice`.

palette

Character vector specifying colors to use for mapping groups to color aesthetic. Should be same length as number of groups. If NULL, default palette will be used. Colors will be mapped to groups based on order of group levels.

pdf

a logical value indicating whether to save the plots as PDF files. Defaults to TRUE.

file.ann

a character string specifying an annotation to add to the file names of the saved plots.

pdf.wid

a numeric value specifying the width of the saved PDF files.

pdf.hei

a numeric value specifying the height of the saved PDF files.

...

further arguments to be passed to the underlying functions.

Value

A list of ggplot2 objects representing the beta ordination plots.

See also

Examples

if (FALSE) {
data(subset_T2D.obj)

generate_beta_ordination_long(
  data.obj = subset_T2D.obj,
  dist.obj = NULL,
  pc.obj = NULL,
  subject.var = "subject_id",
  time.var = "visit_number_num",
  t0.level = NULL,
  ts.levels = NULL,
  group.var = "subject_gender",
  strata.var = "subject_race",
  adj.vars = NULL,
  dist.name = 'BC',
  base.size = 12,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5
)
}