Generate taxa-level difference scatterplot pairs

This function generates scatterplot pairs showing the difference in abundance of taxa at a specified taxonomic level between two time points, with grouping and stratification.

Usage

generate_taxa_change_scatterplot_pair(
  data.obj,
  subject.var,
  time.var,
  group.var = NULL,
  strata.var = NULL,
  change.base = NULL,
  change.func = "difference",
  feature.level = NULL,
  feature.dat.type = c("count", "proportion", "other"),
  features.plot = NULL,
  top.k.plot = NULL,
  top.k.func = NULL,
  prev.filter = 0.01,
  abund.filter = 0.01,
  base.size = 16,
  theme.choice = "bw",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = NULL,
  pdf.wid = 11,
  pdf.hei = 8.5,
  ...
)

Arguments

data.obj

An object containing the OTU table, taxonomy table, and sample metadata.

subject.var

Character string specifying the subject variable in metadata.

time.var

Character string specifying the time variable in metadata.

group.var

Character string specifying the grouping variable in metadata. Default NULL.

strata.var

Character string specifying the stratification variable in metadata. Default NULL.

change.base

Character string specifying the baseline time point.

change.func

Function or character string specifying method to compute change ('difference', 'relative difference', 'lfc'). Default 'difference'.

feature.level

Character vector specifying taxonomic level(s) to include. Default c('Phylum', 'Family', 'Genus').

feature.dat.type

Character string specifying data type, one of "count", "proportion", or "other". Default "count".

features.plot

Character vector specifying specific taxa names to plot. Default NULL.

top.k.plot

Numeric specifying number of top abundant taxa to plot if top.k.func is provided.

top.k.func

Function or character string specifying method to select top abundant taxa (mean, sd, custom function).

prev.filter

Numeric specifying minimum prevalence threshold for filtering taxa.

abund.filter

Numeric specifying minimum abundance threshold for filtering taxa.

base.size

Numeric specifying base font size for plot text elements. Default 16.

theme.choice

Character string specifying ggplot2 theme to use. Default 'bw'.

custom.theme

A ggplot2 theme object to override default theme. Default NULL.

palette

Character vector specifying colors to use for plot. Defaults to RColorBrewer palette.

pdf

Logical, if TRUE, save plot(s) as PDF file(s). Default TRUE.

file.ann

Character string specifying file annotation to add to PDF file name. Default NULL.

pdf.wid, pdf.hei

Numeric specifying PDF width and height in inches. Default 11 x 8.5.

...

Additional arguments passed to ggplot2::ggsave().

Value

A ggplot object containing the scatterplot pair(s).

Examples

if (FALSE) {
# Load required libraries and data
library(vegan)
data(peerj32.obj)
peerj32.obj$meta.dat <- peerj32.obj$meta.dat %>%
dplyr::select(all_of("subject")) %>% dplyr::distinct() %>%
dplyr::mutate(cons = runif(dplyr::n(),0,5)) %>%
dplyr::left_join(peerj32.obj$meta.dat %>% rownames_to_column("sample"),by = "subject") %>%
tibble::column_to_rownames("sample")
# Generate the boxplot pair
plot_list_all <- generate_taxa_change_scatterplot_pair(
  data.obj = peerj32.obj,
  subject.var = "subject",
  time.var = "time",
  group.var = "cons",
  strata.var = "sex",
  change.base = "1",
  change.func = "lfc",
  feature.level = "Genus",
  feature.dat.type = "other",
  features.plot = NULL,
  top.k.plot = 8,
  top.k.func = "mean",
  prev.filter = 0.01,
  abund.filter = 0.01,
  base.size = 16,
  theme.choice = "classic",
  custom.theme = NULL,
  palette = NULL,
  pdf = TRUE,
  file.ann = "test",
  pdf.wid = 11,
  pdf.hei = 8.5
)
}