Compute and Analyze Taxa Changes Over Time

This function from the MicrobiomeStat package computes the taxa changes over time for different groups. The changes are calculated based on a specified function. The function also performs linear modelling on these changes using a specified group variable and covariates, and returns a report summarizing these analyses. If no time variable is provided, the function will select the first unique value from the metadata as the baseline for comparison.

Usage

generate_taxa_change_test_pair(
  data.obj,
  subject.var,
  time.var = NULL,
  group.var,
  adj.vars,
  change.base,
  change.func,
  feature.level,
  prev.filter = 0,
  abund.filter = 0,
  feature.dat.type = c("count", "proportion", "other"),
  ...
)

Arguments

data.obj

A list object in MicrobiomeStat format.

subject.var

The name of the subject variable column in the metadata.

time.var

The name of the time variable column in the metadata (optional).

group.var

The name of the grouping variable column for linear modeling in the metadata.

adj.vars

Names of additional variables to be used as covariates in the analysis.

change.base

The baseline time point for detecting changes in taxa. If NULL, the first unique value from the time.var column will be used (optional).

change.func

The function to be used for calculating changes. Options include: "lfc" (log fold change), "relative difference", and a user-defined function (default is "lfc").

feature.level

The taxonomic level at which to perform the analysis.

prev.filter

A numeric value indicating the minimum prevalence filter for the taxa (default is 0).

abund.filter

A numeric value indicating the minimum abundance filter for the taxa (default is 0).

feature.dat.type

Type of the feature data. Options include: "count", "proportion", and "other" (default is "count").

...

Additional parameters to be passed to the ZicoSeq function.

Value

A named list where each element corresponds to a feature level and contains a dataframe with the calculated taxa changes, their corresponding p-values, and other statistics from the linear model.

Examples

if (FALSE) {
library(vegan)
library(GUniFrac)
library(ape)
library(philentropy)
library(MicrobiomeStat)
data(peerj32.obj)

generate_taxa_change_test_pair(
  data.obj = peerj32.obj,
  subject.var = "subject",
  time.var = "time",
  group.var = "group",
  adj.vars = c("sex"),
  change.base = "1",
  change.func = "lfc",
  feature.level = "original",
  prev.filter = 0.01,
  abund.filter = 0.01,
  feature.dat.type = "count"
)
}